added abilities checking to all routes, tested all routes, modified category routes,added photoURL to author and book models

+++ routes/authors.js
@@ -5,6 +5,9 @@ const createError = require('http-errors');
 
 //add author
 router.post('/', (req, res, next) => {
+    const user = req.user;
+    if (user.Abilities.cannot('add', 'author'))
+        next(createError(401, 'request denied'));
     Author
         .create(req.body)
         .then(user => res.send(user))
@@ -14,6 +17,9 @@ router.post('/', (req, res, next) => {
 //get all authors
 router.get('/', async function (req, res, next) {
     try {
+        const user = req.user;
+        if (user.Abilities.cannot('getAll', 'author'))
+            return next(createError(401, 'request denied'));
         const authors = await Author.find()
             .populate('category', 'name');
         res.send(authors);
@@ -25,6 +31,9 @@ router.get('/', async function (req, res, next) {
 //get author by id
 router.get('/:id', async function (req, res, next) {
     try {
+        const user = req.user;
+        if (user.Abilities.cannot('getById', 'author'))
+            return next(createError(401, 'request denied'));
         const authors = await Author.findById(req.params.id)
             .populate('category', 'name');
         res.send(authors);
@@ -35,6 +44,9 @@ router.get('/:id', async function (req, res, next) {
 
 //delete author
 router.delete('/:id', (req, res, next) => {
+    const user = req.user;
+    if (user.Abilities.cannot('delete', 'author'))
+        return next(createError(401, 'request denied'));
     Author
         .findByIdAndDelete(req.params.id)
         .exec()
@@ -45,6 +57,9 @@ router.delete('/:id', (req, res, next) => {
 //edit author
 router.patch('/:id', (req, res, next) => {
     //debugger;
+    const user = req.user;
+    if (user.Abilities.cannot('update', 'author'))
+        return next(createError(401, 'request denied'));
     Author
         .findByIdAndUpdate(req.params.id, req.body, { new: true })
         .exec()
